Create AzureVM
Creates a Virtual Machine on Microsoft Azure - In this example, the provider block specifies the Azure provider version and features. The azurerm_resource_group block creates a resource group, and the azurerm_virtual_machine block creates a virtual machine within that resource group. The virtual machine is configured with storage, OS disk, OS profile, and OS profile Linux Configuration.

Please be aware that this is just a simple example and those terraform codes will not run if you don't have the correct credentials set up and azure subscription ready. It also may not be in complete form for running the infrastructure.